synchronize Redis proxy traits for different versions
