+++ src/Symfony/Component/Cache/Traits/RedisCluster6Proxy.php
@@ -28,6 +28,7 @@ class RedisCluster6Proxy extends \RedisCluster implements ResetInterface, LazyOb
     use LazyProxyTrait {
         resetLazyObject as reset;
     }
+    use RedisCluster6ProxyTrait;
 
     private const LAZY_OBJECT_PROPERTY_SCOPES = [];
 
@@ -94,11 +95,6 @@ class RedisCluster6Proxy extends \RedisCluster implements ResetInterface, LazyOb
     public function bgrewriteaof($key_or_address): \RedisCluster|bool
     {
         return ($this->lazyObjectState->realInstance ??= ($this->lazyObjectState->initializer)())->bgrewriteaof(...\func_get_args());
-    }
-
-    public function waitaof($key_or_address, $numlocal, $numreplicas, $timeout): \RedisCluster|array|false
-    {
-        return ($this->lazyObjectState->realInstance ??= ($this->lazyObjectState->initializer)())->waitaof(...\func_get_args());
     }
 
     public function bgsave($key_or_address): \RedisCluster|bool
@@ -659,11 +655,6 @@ class RedisCluster6Proxy extends \RedisCluster implements ResetInterface, LazyOb
     public function pttl($key): \RedisCluster|false|int
     {
         return ($this->lazyObjectState->realInstance ??= ($this->lazyObjectState->initializer)())->pttl(...\func_get_args());
-    }
-
-    public function publish($channel, $message): \RedisCluster|bool
-    {
-        return ($this->lazyObjectState->realInstance ??= ($this->lazyObjectState->initializer)())->publish(...\func_get_args());
     }
 
     public function pubsub($key_or_address, ...$values): mixed
